Initial Code Barf
Woah that's a bunch of code right there what the.
So this is groundwork being done to add *buttons* or as the game calls
them *gizmos* to a Jedi character so that they have force powers.
There's also a bit of code to set up a Force Pool to keep track of a
Jedi's available power reserves.