Functional Abilities (Working: Force Heal for Jedi characters)
So in this update, I finalized a lot of what I set out to do with the
ablities framework.

I also created derived classes of the ablities framework that keep in
mind values / entries that are specific to force wielders.

I created two basic spells to play with -- Heal and Drain. Drain is not
working properly at the moment, but it does cast and cause pain. I'm
working on allowinig double verbs on an ability to allow for both the
transference of pain as well as health for a dark force user.